# videodna-studio/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VideoDNA Studio API",
    "description": "API do VideoDNA Studio — análise, desmontagem semântica, edição estrutural e reconstrução de vídeos com múltiplos modelos de IA. Erros seguem o formato `{error: {code, message, details, requestId}}` com códigos normalizados.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/analyze": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analisar o vídeo (idempotente)",
        "operationId": "analyze_projects__project_id__analyze_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AnalyzeRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/analysis": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Status da análise atual",
        "operationId": "get_analysis_projects__project_id__analysis_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/video-dna": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Video DNA (original ou atual = original + operações)",
        "operationId": "get_video_dna_projects__project_id__video_dna_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "original",
                "current"
              ],
              "type": "string",
              "default": "current",
              "title": "View"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDNAResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/entities": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Elementos detectados (personagens, objetos, cenário...)",
        "operationId": "list_entities_projects__project_id__entities_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EntityType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "needsReview",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Needsreview"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityOut"
                  },
                  "title": "Response List Entities Projects  Project Id  Entities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{entity_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Detalhe do elemento",
        "operationId": "get_entity_entities__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "analysis"
        ],
        "summary": "Corrigir a análise de um elemento (vira uma operação CORRECT)",
        "operationId": "correct_entity_entities__entity_id__patch",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityCorrection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "pipelineVersion": {
            "type": "string",
            "title": "Pipelineversion"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "lowConfidenceCount": {
            "type": "integer",
            "title": "Lowconfidencecount",
            "default": 0
          },
          "reusedFromId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reusedfromid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "pipelineVersion",
          "mock",
          "summary",
          "lowConfidenceCount",
          "reusedFromId",
          "createdAt",
          "completedAt"
        ],
        "title": "AnalysisOut"
      },
...
      "AnalysisStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "AnalysisStatus"
      },
      "AnalyzeRequest": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "AnalyzeRequest"
      },
...
      "EntityCorrection": {
        "properties": {
          "property": {
            "type": "string",
            "enum": [
              "label",
              "type",
              "subtype",
              "importance",
              "description"
            ],
            "title": "Property"
          },
          "value": {
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "property",
          "value"
        ],
        "title": "EntityCorrection",
        "description": "Manual correction of an uncertain analysis (\"Copo / Taça?\")."
      },
...
      "EntityOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "subtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtype"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "parentKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentkey"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "importance": {
            "$ref": "#/components/schemas/Importance"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable"
          },
          "replaceable": {
            "type": "boolean",
            "title": "Replaceable"
          },
          "needsReview": {
            "type": "boolean",
            "title": "Needsreview"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/EntityAlternative"
            },
            "type": "array",
            "title": "Alternatives"
          },
          "appearances": {
            "items": {
              "$ref": "#/components/schemas/Appearance"
            },
            "type": "array",
            "title": "Appearances"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "current": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Entity"
              },
              {
                "type": "null"
              }
            ]
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/EditCategoryOut"
            },
            "type": "array",
            "title": "Categories"
          },
          "quickActions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quickactions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "type",
          "subtype",
          "label",
          "description",
          "parentKey",
          "confidence",
          "importance",
          "editable",
          "replaceable",
          "needsReview",
          "attributes",
          "alternatives",
          "appearances",
          "sources",
          "current",
          "categories",
          "quickActions"
        ],
        "title": "EntityOut"
      },
...
      "EntityType": {
        "type": "string",
        "enum": [
          "character",
          "hair",
          "wardrobe",
          "accessory",
          "object",
          "furniture",
          "environment",
          "environment_part",
          "lighting",
          "text",
          "audio"
        ],
        "title": "EntityType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Importance": {
        "type": "string",
        "enum": [
          "ESSENTIAL",
          "IMPORTANT",
          "DECORATIVE"
        ],
        "title": "Importance"
      },
...
      "JobKind": {
        "type": "string",
        "enum": [
          "INGEST",
          "ANALYSIS",
          "PREVIEW",
          "GENERATION"
        ],
        "title": "JobKind"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "title": "Projectid"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0.0
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "planId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planid"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "projectId",
          "kind",
          "status",
          "stage",
          "progress",
          "message",
          "errorCode",
          "errorMessage",
          "result",
          "planId",
          "attempts",
          "createdAt",
          "startedAt",
          "finishedAt"
        ],
        "title": "JobOut"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "QUEUED",
          "PREPARING",
          "ANALYZING",
          "GENERATING",
          "QA",
          "REPAIRING",
          "ASSEMBLING",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "JobStatus"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 8,
            "title": "Password"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "tokenType": {
            "type": "string",
            "const": "bearer",
            "title": "Tokentype",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "tokenType",
          "user"
        ],
        "title": "TokenResponse"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "isAdmin": {
            "type": "boolean",
            "title": "Isadmin",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "displayName",
          "isAdmin"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VideoDNAResponse": {
        "properties": {
          "analysisId": {
            "type": "string",
            "format": "uuid",
            "title": "Analysisid"
          },
          "view": {
            "type": "string",
            "enum": [
              "original",
              "current"
            ],
            "title": "View"
          },
          "editCount": {
            "type": "integer",
            "title": "Editcount"
          },
          "dna": {
            "$ref": "#/components/schemas/VideoDNA"
          },
          "assetUrls": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Asseturls"
          }
        },
        "type": "object",
        "required": [
          "analysisId",
          "view",
          "editCount",
          "dna",
          "assetUrls"
        ],
        "title": "VideoDNAResponse"
      }
    }
  },
...
}